refactor(benchmarks): split cli.py into a typer sub-app package
The 933-line cli.py was a god-file whose ``# --- section ---`` banners
signalled it wanted to be separate modules. Split one module per command
group, all registering onto a shared ``app`` so the flat command surface
and ``--help`` order stay identical:

  _base.py        app/memory_app, shared types + helpers
  introspect.py   list / show / filter
  run.py          smoke / run / notebook
  sweep.py        sweep
  compare.py      compare
  plot.py         plot
  memory.py       memory save / sweep / compare

Also tidy two over-narrated spots:
- sweep.py: drop a stale comment that claimed ``cwd`` was pinned to the
  repo root when the code runs ``cwd=import_dir``; trim the isolation
  blocks the function docstring already covers in full.
- plotting.py: tighten the facet-layout comment.

CLI surface, ruff, and mypy all verified unchanged.
